Add login  and validation in REST

+++ app/Http/Controllers/API/TaskController.php
@@ -11,42 +11,63 @@ use App\Http\Resources\Task as TaskResource;
 class TaskController extends Controller
 {
     public function create(TaskRequest $request){
-       Task::create($request->all());
-       return response()->json([
-           'status'=>201
-       ]);
+        $user=auth()->user();
+
+        Task::create([
+            'date'=>$request->get('date'),
+            'description'=>$request->get('description'),
+            'user_id'=>$user->id
+        ]);
+        return response()->json([
+            'status'=>201
+        ],201);
    }
 
    public function index(){
-        $id=1;
-        return TaskResource::collection(Task::where('user_id',$id)->get());
+        $user=auth()->user();
+        return TaskResource::collection(Task::where('user_id',$user->id)->get());
    }
 
    public function update(TaskRequest $request, $id){
+       $user=auth()->user();
+
        $task=Task::find($id);
 
        if($task==null){
-           return response()->json(['status'=>404]);
+           return response()->json(['status'=>404],404);
        }
 
-       $task->update($request->all());
+       if($task->user_id!=$user->id){
+           return response()->json(['status'=>401],401);
+       }
+
+       $task->update([
+           'date'=>$request->get('date'),
+           'description'=>$request->get('description')
+       ]);
+
        return response()->json([
            'status'=>200
-       ]);
+       ],200);
    }
 
     public function delete( $id){
+        $user=auth()->user();
         $task=Task::find($id);
 
         if($task==null){
-            return response()->json(['status'=>404]);
+            return response()->json(['status'=>404],404);
+        }
+
+        if($task->user_id!=$user->id){
+            return response()->json(['status'=>401],401);
         }
 
         $task->delete();
 
         return response()->json([
             'status'=>200
-        ]);
+        ],200);
     }
